refactor(mcp): split HTTP dispatch from stdio session and 1k files
HTTP and stdio no longer share dummy session state. Protocol and tools
are separate modules. POST /mcp is its own mw-serve path. Rho MCP merge
applies a target record. The capture hook no longer loads client auth.
